refactor(breach-history-extract): align with repo conventions
/simplify review of the extraction surfaced three real issues:

- Directory naming: `_components/` is a convention used under `app/<route>/`,
  not under `components/`. The existing precedent for grouping related
  shared components is a named subdir (e.g. `pool-header/`). Move the two
  new files into `components/breach-history/` so the layout matches and
  signals "these belong with breach-history-panel".
- `ONE_HOUR` / `ONE_DAY` were new constants in `bucket-filter.tsx`, but
  `lib/time-series.ts` already exports `SECONDS_PER_DAY` for the homepage
  charts. Add `SECONDS_PER_HOUR` there and import both from time-series in
  the panel — one shared time-constant module instead of three.
- `DurationField` was exported even though only `DurationRangeInputs`
  (its parent in the same module) consumes it. Drop the export to keep
  the module surface minimal.

51 characterization tests still green.
